refactor(eventDelegators) : make eventDelegators
